refactor manage documents to manage news articles
