+++ DfE.GIAP.All/src/DfE.GIAP.Web/ViewModels/Admin/ManageDocumentsViewModel.cs
@@ -11,7 +11,6 @@ public class ManageDocumentsViewModel
     public Document DocumentList { get; set; }
     public CommonResponseBodyViewModel DocumentData { get; set; }
     public string SelectedNewsId { get; set; }
-    public bool HasInvalidDocumentList { get; set; }
     public bool HasInvalidNewsList { get; set; }
     public Confirmation Confirmation { get; set; }
     public NewsArticleViewModel NewsArticle { get; set; }
