+++ DfE.GIAP.All/src/DfE.GIAP.Web/Views/Admin/Index.cshtml
@@ -47,9 +47,9 @@
                         @if (@Model.IsAdmin)
                         {
                             <div class="govuk-radios__item">
-                                <input class="govuk-radios__input" id="manage-documents" asp-for="SelectedAdminOption" type="radio" value="ManageDocuments">
+                                <input class="govuk-radios__input" id="manage-documents" asp-for="SelectedAdminOption" type="radio" value="ManageNewsArticles">
                                 <label class="govuk-label govuk-radios__label" for="manage-documents">
-                                    Manage documents
+                                    Manage news articles
                                 </label>
                             </div>
                         }
