--- DfE.GIAP.All/src/DfE.GIAP.Web/Views/Admin/ManageDocuments/ManageDocuments.cshtml
@@ -1,58 +1,0 @@
-﻿@model ManageDocumentsViewModel
-@{
-    ViewBag.Title = "Manage documents";
-}
-
-
-<div class="govuk-width-container">
-    <partial name="Layout/_LayoutHeaderBackButton" model="@Model.BackButton" />
-
-    <h1 class="govuk-heading-xl">Manage documents</h1>
-
-    <form method="post">
-        <div class="govuk-form-group">
-
-            <label for="select-document-dropdown" class="govuk-label">Please select a document</label>
-
-            <select id="select-document-dropdown" asp-for="@Model.DocumentList.DocumentId" class="govuk-select @(Model.HasInvalidDocumentList ? "govuk-textarea--error" : "")" asp-items="ViewBag.ListOfDocuments">
-            </select>
-            <button asp-controller="ManageDocuments" asp-action="ManageDocuments" id="select-document--button" value="Select" name="select" class="govuk-button">Select</button>
-
-            @if (Model?.DocumentList?.DocumentId == DocumentType.Article.ToString() && ViewBag.NewsDocuments != null)
-            {
-                <br />
-
-                <label for="select-article-dropdown" class="govuk-label">Please select a news article</label>
-                <select id="select-article-dropdown" asp-for="@Model.SelectedNewsId" class="govuk-select @(Model.HasInvalidNewsList ? "govuk-textarea--error" : "")" asp-items="ViewBag.NewsDocuments">
-                </select>
-                <button asp-controller="ManageDocuments" asp-action="SelectNewsArticle" id="select-document--button" value="Select" name="select" class="govuk-button">Select</button>
-            }
-
-
-            @if (Model?.DocumentList?.DocumentId == DocumentType.Article.ToString())
-            {
-                <p class="govuk-body">
-                    <a asp-controller="ManageDocuments" asp-action="CreateNewsArticle" id="create-new-article--button" value="Create" name="create" class="govuk-button govuk-button--secondary">Create new article</a>
-                </p>
-            }
-
-
-            @if (Model?.DocumentList?.DocumentId != DocumentType.Article.ToString() &&
-                        Model?.DocumentData is not null)
-            {
-                <h2 class="govuk-heading-l">Edit document</h2>
-
-                <input type="hidden" asp-for="@Model.DocumentData.Id" />
-                <input type="hidden" asp-for="@Model.DocumentData.Published" />
-
-                <br />
-                <partial name="~/Views/Admin/ManageDocuments/_DocumentFields.cshtml" for="DocumentData" />
-
-                <div class="govuk-button-group">
-                    <button asp-controller="ManageDocuments" asp-action="PublishChanges" id="publish-document--button" name="publish" value="Publish" class="govuk-button govuk-button--secondary">Publish</button>
-                    <button asp-controller="ManageDocuments" asp-action="PreviewChanges" id="preview-document--button" name="preview" value="Preview" class="govuk-button">Preview</button>
-                </div>
-            }
-        </div>
-    </form>
-</div>
