+++ DfE.GIAP.All/src/DfE.GIAP.Web/Views/Admin/ManageNewsArticles/EditNewsArticle.cshtml
@@ -60,9 +60,9 @@
                 </div>
 
                 <div class="govuk-button-group">
-                    <a asp-controller="ManageDocuments" asp-action="ManageDocuments" id="discard-news-article--button" name="discard" value="Discard" class="govuk-button govuk-button--secondary">Discard</a>
-                    <button asp-controller="ManageDocuments" asp-action="DeleteNewsArticle" id="delete-document--button" name="delete" value="Delete" type="submit" class="govuk-button govuk-button--secondary">Delete</button>
-                    <button asp-controller="ManageDocuments" asp-action="UpdateNewsArticle" id="update-document--button" name="update" value="Update" type="submit" class="govuk-button">Update</button>
+                    <a asp-controller="ManageNewsArticles" asp-action="ManageNewsArticles" id="discard-news-article--button" name="discard" value="Discard" class="govuk-button govuk-button--secondary">Discard</a>
+                    <button asp-controller="ManageNewsArticles" asp-action="DeleteNewsArticle" id="delete-document--button" name="delete" value="Delete" type="submit" class="govuk-button govuk-button--secondary">Delete</button>
+                    <button asp-controller="ManageNewsArticles" asp-action="UpdateNewsArticle" id="update-document--button" name="update" value="Update" type="submit" class="govuk-button">Update</button>
                 </div>
             </div>
         </form>
