+++ DfE.GIAP.All/src/DfE.GIAP.Web/Views/Admin/ManageNewsArticles/Error.cshtml
@@ -1,6 +1,6 @@
 ﻿@model UserErrorViewModel
 @{
-    ViewData["Title"] = "Error: There is a problem with documents";
+    ViewData["Title"] = "Error: There is a problem";
 }
 
 <div class="govuk-grid-row">
